+++ loop-templates/templates/src/aptitudes/index.ts
@@ -6,6 +6,7 @@ import network from './network';
 import process from './process';
 import ui from './ui';
 import user from './user';
+import vault from './vault';
 import window from './window';
 import index from './index.ts.squirrelly';
 import { FileMap } from '@/types';
@@ -19,6 +20,7 @@ const fileMap: FileMap = {
   process: { fileName: 'process', aptitude: 'process' },
   ui: { fileName: 'ui', aptitude: 'any' },
   user: { fileName: 'user', aptitude: 'user' },
+  vault: { fileName: 'vault', aptitude: 'vault' },
   window: { fileName: 'window', aptitude: 'window' },
   index: { fileName: 'index.ts', aptitude: 'any' },
 };
@@ -32,6 +34,7 @@ export default {
   process,
   ui,
   user,
+  vault,
   window,
   index,
   fileMap,