+++ vscode-extension/CHANGELOG.md
@@ -7,6 +7,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+- Added Vault aptitude to VSCode/npx tools.
+
 ### Changed
 
 ## [0.4.0]